annotation_project/requirements.txt
# Django & REST
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# DB
psycopg2-binary==2.9.9

# Excel / data import
pandas==2.2.2
openpyxl==3.1.5

# CORS
django-cors-headers==4.3.1